Refactor constants and types for improved organization and clarity in Login and Register components
