+++ src/shared/constants/Global.constants.ts
@@ -6,11 +6,3 @@ export const DASHBOARD_ROUTE = '/dashboard';
 export const REGISTER_ROUTE = '/register';
 export const FORGOT_PASSWORD_ROUTE = '/forgot-password';
 export const LOGIN_ROUTE = '/';
-
-// Messaging
-export const SUCCESS_CREATE_USER_TITLE = '¡Bienvenido!';
-export const SUCCESS_CREATE_USER_SPAN = 'Tu cuenta ya está lista'
-export const SUCCESS_CREATE_USER_MESSAGE = '¡Vamos a por esas metas financieras!';
-export const SUCESS_CREATE_USER_SECONDARY_MESSAGE = 'Inicia sesión y empieza a usar Budget Master.'
-export const ERROR_CREATE_USER_TITLE = 'Oops! Algo no salió como esperabamos.';
-export const ERROR_CREATE_USER_MESSAGE = 'Por favor intente nuevamente más tarde.';
