+++ src/shared/constants/Login.constants.ts
@@ -1,1 +1,16 @@
 export const ERROR_EMAIL_IN_USE = 'Try with other email.'
+
+// Errors
+export const ERROR_UNAUTHORIZED_LOGIN = 'Email or Password incorrect.'
+export const ERROR_UNAUTHORIZED_LOGIN_MESSAGE = 'Correo electronico o contraseña incorrecta.';
+export const ERROR_PASSWORD_REQUIRED = 'Contraseña es requerida'
+export const ERROR_EMAIL_REQUIRED = 'Por favor, ingrese su correo electrónico';
+export const ERROR_INVALID_EMAIL = 'Correo electrónico inválido';
+export const ERROR_CREATE_USER_TITLE = 'Oops! Algo no salió como esperabamos.';
+export const ERROR_CREATE_USER_MESSAGE = 'Por favor intente nuevamente más tarde.';
+
+// Messaging
+export const SUCCESS_CREATE_USER_TITLE = '¡Bienvenido!';
+export const SUCCESS_CREATE_USER_SPAN = 'Tu cuenta ya está lista'
+export const SUCCESS_CREATE_USER_MESSAGE = '¡Vamos a por esas metas financieras!';
+export const SUCESS_CREATE_USER_SECONDARY_MESSAGE = 'Inicia sesión y empieza a usar Budget Master.'
